pyproject.toml
[project]
name = "evilve"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "fastapi>=0.128.1",
    "langsmith[openai-agents]>=0.6.9",
    "loguru>=0.7.3",
    "mlflow>=3.9.0",
    "openai-agents>=0.7.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyyaml>=6.0.2",
    "sqlmodel>=0.0.32",
    "uvicorn>=0.40.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.7",
]
